requirements/base.txt
django==1.11
django-environ==0.4.4
djangorestframework==3.7.7